+++ packages/HuYaRecorder/src/huya_mobile_api.ts
@@ -4,7 +4,9 @@
 import axios from "axios";
 
 import { utils } from "@bililive-tools/manager";
-import { assert } from "./utils.js";
+import { assert, getFormatSources } from "./utils.js";
+
+import type { StreamResult, StreamProfile } from "./types.js";
 
 const requester = axios.create({
   timeout: 10e3,
@@ -16,7 +18,7 @@ const requester = axios.create({
 
 export async function getRoomInfo(
   roomIdOrShortId: string,
-  formatName: "auto" | "flv" | "hls" = "auto",
+  formatPriorities: Array<"flv" | "hls"> = ["flv", "hls"],
 ) {
   const res = await requester.get(
     `https://mp.huya.com/cache.php?m=Live&do=profileRoom&roomid=${roomIdOrShortId}`,
@@ -70,6 +72,11 @@ export async function getRoomInfo(
   };
 
   const startTime = new Date(profile.liveData?.startTime * 1000);
+  const formatSources = getFormatSources(sources, formatPriorities);
+  if (!formatSources) {
+    throw new Error("No format sources found");
+  }
+
   return {
     living: profile.liveStatus === "ON",
     id: profile.liveData.profileRoom,
@@ -78,8 +85,8 @@ export async function getRoomInfo(
     roomId: profile.liveData.profileRoom,
     avatar: profile.liveData.avatar180,
     cover: profile.liveData.screenshot,
-    streams: formatName === "hls" ? streams.hls : streams.flv,
-    sources: formatName === "hls" ? sources.hls : sources.flv,
+    streams: formatSources.formatName === "hls" ? streams.hls : streams.flv,
+    sources: formatSources.sources,
     startTime,
     liveId: utils.md5(`${roomIdOrShortId}-${startTime?.getTime()}`),
   };
@@ -250,24 +257,3 @@ const cdn = {
   HS: "火山",
   WS: "网宿",
 } as const;
-
-interface StreamResult {
-  flv: {
-    name: string;
-    url: string;
-  }[];
-  hls: {
-    name: string;
-    url: string;
-  }[];
-}
-
-export interface StreamProfile {
-  desc: string;
-  bitRate: number;
-}
-
-export interface SourceProfile {
-  name: string;
-  url: string;
-}